+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [v0.1.3]
+
+### Added
+
+- `vector_matching` check for finding queries with incorrect `on()` or `ignoring()`
+  keywords.
+
+### Fixed
+
+- `comparison` check would trigger false positive for rules using `unless` keyword.
+
 ## [v0.1.2]
 
 ### Fixed